9月4号更新

+++ src/main/java/llcweb/com/domain/models/Document.java
@@ -17,9 +17,9 @@ public  class Document {
    //文章id
    @Id
    @GeneratedValue
-   private long id;
+   private Integer id;
    //作者id
-   private int authorId = 0;
+   private Integer authorId = 0;
    //作者
    private String author;
    //文章标题
@@ -49,20 +49,18 @@ public  class Document {
       this.infor = infor;
       this.model = model;
    }
-
-   public long getId() {
+   public Integer getId() {
       return id;
    }
-
-   public void setId(long id) {
+   public void setId(Integer id) {
       this.id = id;
    }
 
-   public int getAuthorId() {
+   public Integer getAuthorId() {
       return authorId;
    }
 
-   public void setAuthorId(int authorId) {
+   public void setAuthorId(Integer authorId) {
       this.authorId = authorId;
    }
